+++ pyogrio/tests/test_geopandas_io.py
@@ -1,5 +1,6 @@
 import contextlib
 import locale
+import os
 import warnings
 from datetime import datetime
 from io import BytesIO
@@ -91,6 +92,17 @@ def skip_if_no_arrow_write_api(request):
         and request.node.get_closest_marker("requires_arrow_write_api")
     ):
         pytest.skip("GDAL>=3.8 required for Arrow write API")
+
+
+@contextlib.contextmanager
+def use_arrow_context():
+    original = os.environ.get("PYOGRIO_USE_ARROW", None)
+    os.environ["PYOGRIO_USE_ARROW"] = "1"
+    yield
+    if original:
+        os.environ["PYOGRIO_USE_ARROW"] = original
+    else:
+        del os.environ["PYOGRIO_USE_ARROW"]
 
 
 def spatialite_available(path):
@@ -372,6 +384,48 @@ def test_read_datetime_tz(datetime_tz_file, tmp_path, use_arrow):
         # with Arrow, the datetimes are always read as UTC
         expected = expected.dt.tz_convert("UTC")
     assert_series_equal(df_read.datetime_col, expected)
+
+
+def test_read_list_types(list_field_values_file, use_arrow):
+    if not GDAL_GE_352:
+        pytest.xfail(reason="GDAL 3.4.3 didn't handle all list types perfectly")
+
+    result = read_dataframe(list_field_values_file, use_arrow=use_arrow)
+
+    assert "list_int64" in result.columns
+    assert result["list_int64"][0].tolist() == [0, 1]
+    assert result["list_int64"][1].tolist() == [2, 3]
+    assert result["list_int64"][2].tolist() == []
+    assert result["list_int64"][3] is None
+    assert result["list_int64"][4] is None
+    assert "list_double" in result.columns
+    assert result["list_double"][0].tolist() == [0.0, 1.0]
+    assert result["list_double"][1].tolist() == [2.0, 3.0]
+    assert result["list_double"][2].tolist() == []
+    assert result["list_double"][3] is None
+    assert result["list_double"][4] is None
+    assert "list_string" in result.columns
+    assert result["list_string"][0].tolist() == ["string1", "string2"]
+    assert result["list_string"][1].tolist() == ["string3", "string4", ""]
+    assert result["list_string"][2].tolist() == []
+    assert result["list_string"][3] is None
+    assert result["list_string"][4] == [""]
+
+    # Once any row of a column contains a null value in a list (in the test geojson),
+    # the column isn't recognized as a list column anymore and the values are returned
+    # as strings.
+    assert "list_int_with_null" in result.columns
+    assert result["list_int_with_null"][0] == "[ 0, null ]"
+    assert result["list_int_with_null"][1] == "[ 2, 3 ]"
+    assert result["list_int_with_null"][2] == "[ ]"
+    assert pd.isna(result["list_int_with_null"][3])
+    assert pd.isna(result["list_int_with_null"][4])
+    assert "list_string_with_null" in result.columns
+    assert result["list_string_with_null"][0] == '[ "string1", null ]'
+    assert result["list_string_with_null"][1] == '[ "string3", "string4", "" ]'
+    assert result["list_string_with_null"][2] == "[ ]"
+    assert pd.isna(result["list_string_with_null"][3])
+    assert result["list_string_with_null"][4] == '[ "" ]'
 
 
 @pytest.mark.filterwarnings(
@@ -2212,6 +2266,29 @@ def test_arrow_bool_exception(tmp_path, ext):
 
     else:
         _ = read_dataframe(filename, use_arrow=True)
+
+
+@requires_pyarrow_api
+def test_arrow_enable_with_environment_variable(tmp_path):
+    """Test if arrow can be enabled via an environment variable."""
+    # Latin 1 / Western European
+    encoding = "CP1252"
+    text = "ÿ"
+    test_path = tmp_path / "test.gpkg"
+
+    df = gp.GeoDataFrame({text: [text], "geometry": [Point(0, 0)]}, crs="EPSG:4326")
+    write_dataframe(df, test_path, encoding=encoding)
+
+    # Without arrow, specifying the encoding is supported
+    result = read_dataframe(test_path, encoding="cp1252")
+    assert result is not None
+
+    # With arrow enabled, specifying the encoding is not supported
+    with use_arrow_context():
+        with pytest.raises(
+            ValueError, match="non-UTF-8 encoding is not supported for Arrow"
+        ):
+            _ = read_dataframe(test_path, encoding="cp1252")
 
 
 @pytest.mark.filterwarnings("ignore:File /vsimem:RuntimeWarning")